+++ cmd/pull.go
@@ -15,9 +15,10 @@ var pullCmd = &cobra.Command{
 	Use:   "pull <NAME PATTERN>",
 	Short: "Pull TMs from a catalog.",
 	Long: `Pulls one or more TMs from a catalog by name pattern, filters or search. 
-The pattern can be a full name or just a prefix the names shall start with. 
-Name pattern, filters and search can be combined to narrow down the result.
-Use --exact to force full-length matching of the name`,
+The pattern can be a full name or a prefix the names shall start with. A partial pattern will match only complete 
+path parts. E.g. 'company/device' will not match 'company/device-mark2', but will match 'company/device/submodel'.
+
+Name pattern, filters and search can be combined to narrow down the result.`,
 	Args: cobra.MaximumNArgs(1),
 	Run:  executePull,
 }
@@ -31,7 +32,6 @@ func init() {
 	pullCmd.Flags().StringVar(&pFilterFlags.FilterManufacturer, "filter.manufacturer", "", "filter TMs by one or more comma-separated manufacturers")
 	pullCmd.Flags().StringVar(&pFilterFlags.FilterMpn, "filter.mpn", "", "filter TMs by one or more comma-separated mpn (manufacturer part number)")
 	pullCmd.Flags().StringVarP(&pFilterFlags.Search, "search", "s", "", "search TMs by their content matching the search term")
-	pullCmd.Flags().BoolP("exact", "e", false, "match the TM name exactly. overrides all other search filter flags")
 	_ = pullCmd.MarkFlagRequired("output")
 }
 
@@ -50,12 +50,7 @@ func executePull(cmd *cobra.Command, args []string) {
 	if len(args) > 0 {
 		name = args[0]
 	}
-	exact, err := cmd.Flags().GetBool("exact")
-	if err != nil {
-		cli.Stderrf("invalid --exact flag")
-		os.Exit(1)
-	}
-	search := cli.CreateSearchParamsFromCLI(pFilterFlags, name, exact)
+	search := cli.CreateSearchParamsFromCLI(pFilterFlags, name)
 	err = cli.NewPullExecutor(remotes.DefaultManager()).Pull(spec, search, outputPath)
 
 	if err != nil {